[GD-2485] Implement GAM limited ads variant
This change is rather heavy as it does a few things

1. the gpt.js is now loaded from the ad tag depending on the given consent data
2. fetching consent is natively integrated into the ad tag. The TcData object is passed with the ad pipeline context. This allows all modules to act according to the given consent and don't need to implement their own fetching logic

As a result this requires us and the publisher to change a few things

1. Remove all consent modules
2. Remove the gpt tag from the head

In practice this should have little effect as gpt.js loads rather quickly, but this change may be followed with some improvements, e.g. a parallel prebid pipeline.