Add labels feature to blocklist module
The blocklist module is probably not going to live that much longer, so
we extended its capabilities a bit to solve a direct issue: deactivating
an ad slot on certain urls.

This implementation extends the blocklist to add arbitrary labels based
on href matches. Ad slots can now be labeled and the blocklist can block
them by adding, or not adding the required label.