First draft for a custom interstitial format (#438)
* First draft for a custom interstitial format

* Adding an idea for an interstitial context

Probably not necessary, but a road to explore

* Implement interstitial state machine

This commit also ports a fix from the 4.x branch.
The impressionViewable event is now used for the
frequency capping of the GAM interstitial.

Also adds the format key value to identify the
interstitial channel properly

* Add tests for the interstitial state machine

* Simplify state management

Still figuring out if beforeRequestAds is actually needed

* Add test for format key-value

* Remove unused type

* Tests for ad unit resolving

* Remove beforeRequestAds